change testing from unittest to pytest (#9)
Fix and enable the `test_datastore.py` datastore test

Due to switch to python3, indices which were obtained by / operator will give floats, which is remediated by replacing it with // operator.

+++ datastore/filesystem/test_filesystem.py
@@ -1,6 +1,5 @@
 import os
 import shutil
-import unittest
 
 from datastore.core import serialize
 from datastore.core.test.test_basic import TestDatastore
@@ -16,15 +15,13 @@ class TestFileSystemDatastore(TestDatastore):
             shutil.rmtree(self.tmp)
 
     def tearDown(self):
-        shutil.rmtree(self.tmp)
+        if os.path.exists(self.tmp):
+            shutil.rmtree(self.tmp)
 
     def test_datastore(self):
         dirs = map(str, range(0, 4))
         dirs = map(lambda d: os.path.join(self.tmp, d), dirs)
         fses = map(FileSystemDatastore, dirs)
-        dses = map(serialize.shim, fses)
+        dses = list(map(serialize.shim, fses))
         self.subtest_simple(dses, numelems=500)
 
-
-if __name__ == '__main__':
-    unittest.main()
